Bypass transit parsing when :body-keys option is given
This commit introduces the :body-keys option to wrap-castra. Keys
specified in this option may contain the rpc expression. Then, if the
request contains one of these keys (checked in order), the value for
that key is used for the rpc expression and no deserialization of the
request :body or serialization of the response :body will be done.

Example:

  The wrap-restful-format middleware can be used to read and
  write Transit JSON. Given a request with content type transit+json
  it will parse the :body and add the parsed EDN forms to the :body-params
  key on the request.

    (def app
      (-> routes
          (wrap-castra {:body-keys [:body-params]} 'my.rpc.namespace)
          (wrap-restful-format)))