explicitly set response charset to utf-8.
if the charset is not specified on the Content-Type header, it will
be determined elsewhere.  for example, the tomcat servlet container
will default to ISO-8859-1, while jetty will encode application/json
in UTF-8.  making this explicit (UTF-8 seems like the most sensible
default here) eliminates the potential for the sorts of nasty post-
deployment surprises one might experience when this setting is left
to the whims of an individual servlet container.